+++ CHANGELOG.md
@@ -6,8 +6,10 @@ All notable changes to COMMIT will be documented in this file.
 
 ### Changed
 - trk2dictionary.run(): removed 'gen_trk' option
-- save_results(): removed 'save_coeff' and 'save_opt_details'
-- added 'get_coeffs' function to get all estimated coefficients
+- save_results(): removed 'save_coeff' and 'save_opt_details' parameters
+- save_results(): now saving only streamline_weights.txt (not anymore xic.txt, xec.txt, xiso.txt)
+- save_results(): added 'stat_coeffs' parameter for saving streamline weights
+- added 'get_coeffs()' function to get all estimated coefficients
 
 ## [1.3.9] - 2020-06-09